requirements.txt
# API Framework
fastapi==0.109.0
pydantic==2.6.0
pydantic-settings==2.0.3
python-multipart==0.0.7
email-validator==2.1.0.post1

# ASGI Server
uvicorn==0.28.0

# Database
sqlalchemy==2.0.27
alembic==1.15.1
asyncpg==0.29.0
psycopg2-binary==2.9.9  # For Alembic offline migrations

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.27.0
pytest-cov==4.1.0

# Development
black==25.1.0
isort==5.13.2
flake8==7.0.0
flake8-bugbear==24.2.6
flake8-docstrings==1.7.0
mypy==1.8.0
pre-commit==3.5.0

# Development tools
python-dotenv==1.0.0
requests==2.31.0 